+++ programs/cp-amm/src/tests/test_position_delegate_permission.rs
@@ -5,7 +5,7 @@ use crate::{
 
 #[test]
 fn test_position_with_full_permission() {
-    let permission: u128 = 0b11111111;
+    let permission: u32 = 0b11111111;
     assert!(
         permission > 1 << (MAX_POSITION_DELEGATE_PERMISSION - 1)
             && permission < 1 << MAX_POSITION_DELEGATE_PERMISSION
